lib/aws_inspector/pyproject.toml
[project]
name = "aws-inspector"
version = "1.0.0"
description = "AWS resource discovery and inventory library"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "moto>=5.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aws_inspector"]

[tool.hatch.build]
include = [
    "*.py",
    "core/*.py",
    "services/*.py",
    "inventory/*.py",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
